requirements.txt
aiosqlite==0.20.0
alembic==1.13.3
annotated-types==0.7.0
anyio==4.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
cryptography==43.0.1
databases==0.9.0
dnspython==2.6.1
email-validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.0
fastapi-cli==0.0.5
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.6
httptools==0.6.1
httpx==0.27.2
idna==3.10
importlib-metadata==8.5.0
importlib-resources==6.4.5
itsdangerous==2.2.0
jinja2==3.1.4
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
orjson==3.10.7
pdfminer.six==20231228
pdfplumber==0.11.4
pillow==10.4.0
pycparser==2.22
pydantic==2.9.2
pydantic-core==2.23.4
pydantic-extra-types==2.9.0
pydantic-settings==2.5.2
pygments==2.18.0
pypdfium2==4.30.0
python-dotenv==1.0.1
python-multipart==0.0.12
PyYAML==6.0.2
rich==13.9.1
shellingham==1.5.4
sniffio==1.3.1
SQLAlchemy==2.0.35
sqlmodel==0.0.22
starlette==0.38.6
typer==0.12.5
typing-extensions==4.12.2
ujson==5.10.0
uvicorn==0.31.0
uvloop==0.20.0
watchfiles==0.24.0
websockets==13.1
zipp==3.20.2
